+++ calliope/constraints/ramping.py
@@ -21,7 +21,8 @@ def ramping_rate(model):
         if m.t.order_dict[t] <= 1:
             return cp.Constraint.NoConstraint
         else:
-            diff = m.e[y, x, t] - m.e[y, x, model.prev(t)]
+            carrier = model.get_option(y + '.carrier')
+            diff = m.e[carrier, y, x, t] - m.e[carrier, y, x, model.prev(t)]
             max_ramping_rate = ramping_rate * m.time_res[t] * m.e_cap[y, x]
             if direction == 'up':
                 return diff <= max_ramping_rate