+++ changelog.rst
@@ -5,8 +5,9 @@ Release History
 0.2.0 (dev)
 +++++++++++
 
+* Added cost classes with an new set ``k``
+* Added energy carriers with a new set ``c``
 * Ability to arbitrarily nest model configuration files with ``import`` statements
-* Added cost classes with an associated new index (``k``)
 * Cleared up terminology (nodes vs locations)
 
 0.1.0 (2013-12-10)